Update index.html
Established drop-down feature for meetings and users when adding an action item. Now after a user has added a meeting, when adding a new action item they can select from a populated drop-down menu of meetings. This has also been implemented for users. If a user adds in a new user, when adding an action item a drop-down menu is populated with that added user as well.