Update app.py
When adding an action item, a checkbox input type has been added in. When the checkbox is toggled, it is picked up by the route handler to mark this action item as completed. If it is not marked as complete (left unchecked), it will be stored as "not complete". 

I also present a handling route for when the user is adding a new user to the database.