Create add_action_item.html
A separate static page for adding action items once a meeting and a user has been added. This follows the flow of allowing a user to visit the tool's homepage and insert a meeting associated with the action item, as well as the user that is responsible for the action item.

Then, there is a button that once clicked will direct the user to a new page where the user can now add an action item. The action item has a drop-down for both meetings and users. The user can only add an action item based off of the available meetings and users added in the prior page (as well as previous entries saved in the underlying database). Once added, this directs back to the homepage index.