Update index.html
What was add_action_item.html is now index.html. Why? The primary purpose of the application is to provide a quick tool for allowing a user to add an action item. This assume that meetings and users have already been added (and thus adding an action item is a matter of selecting from a drop-down menu of already added meetings and users). Unless the meeting and user has not been added, this is our landing page. Also allows for navigating to a "settings" page where new users and meetings can be added.