Update dml.py
For adding a meeting, removed the date argument so it only takes meeting name now. The date argument is added to add_action_item() allowing for a DATETIME attribute to be added when an action item is added. This works well with SQLite. This was a design decision, when an action item is added I want the date and time to be added automatically and stored. I don't need the date of the meeting.