Update dml.py
New function that allows for changing the database field for 'completion' from incomplete to complete. This is an update operation that allows for a user to edit the completion status of an action item. Once an action item has been completed, we can remove it from the action items that are still to be completed.