Update index.html
Presents action items in a table. Currently working on text-wrapping the table. Includes button to mark an action item as complete, once the form /mark-complete is submitted, a list of action item id's which are integers are passed as a string to the server. Returned is a refreshed page that has the action items removed that are no longer marked as not completed.