Update dml.py
New functions added include get_id_by_action_item_name which returns the action_item_id from the request.form['action_item_name'], allowing for displaying the names of the action items in the underlying database to the user but not the id. By getting the id, I am able to then ensure that the new_note will get added to the correct action item. append_note_to_action_item does exactly that updating the Notes column. get_timestamp() allows for accessing the current local time in a formatted way using the time module import in python. I can then add to the Date column the current formatted time so that the user can view when the action item was edited with the new note. This allows for seeing all historical edits to the action item.