Update index.html
Moved the ability to view action items to a new page entirely where the soul purpose was for viewing and interacting with the action items. Thus, the landing page becomes more of a navigating page to the different features the web application has to offer. It also still contains the ability to add an action item.