Update app.py
When an action item is added through the index route, the init_note_action_item() function is called. When a note is appended to an action item, the route add_note adds a timestamp as well as the note. Allowing for organized notes in the viewing front-end. action_items_page returns a new parameter called completed_action_items which calls the get_all_completed_action_items() which is important for displaying the new table being implemented in action_items.html for Completed Action Items so a user of the web application can now view which action items are complete and are not complete all on the same page.