Update action_items.html
New table displaying completed action items. After a user of the web application has marked the action item as completed, this immediately will be placed in the completed action items table. This allows for someone to see what has been completed and what is still in-progress. I changed the formatting and organization of this page, the size of the headers emphasizing "In-Progress Action Items" and "Completed Action Items", which are the two tables.