Update app.py
filter_completed_by_meeting handles the route for when a user of the application wants to filter the presented action items based on meeting, that have also been marked as completed. filter_completed_by_user() handles for filtering action items based off of the completion status set to completed as well as the input of which user name to sort for.