Update action_items.html
New labels for drop-down menus of filtering by the meeting name and then again for filtering by the user name. The form accesses the first index of each, which corresponds to the id of the meeting and user. On the front-end, this is translated into presenting the meeting name and user name which is more readable than the integer datatype of the id that is being stored in the database entities.