Update app.py
This week, my code was moved to a hosting server platform called PythonAnywhere. This involved editing the WSGI file that was provided to ensure my database schema would be properly initialized. Here you can see that I am initializing the database file through my app.py, this is new. I use to just run python3 ddl.py each time I wanted to create the database schema. This checks to see if the database file has already been created, if it exists within the file path. If it has not, then it does not need to be initialized.